py-app/requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.1.3
packaging==24.2
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
ruff==0.8.1
safetensors==0.4.5
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
